requirements.txt
Flask
requests
python-decouple
python-dotenv
flask_cors